Attempting to fix no 201 on post
So far, part of the failure was the Post would return a 400 (bad request) because the default value to post was 0 which fails the initial test conditions.

I am still working on the lack of a 201 response. It seems that it is not getting the 201 in the response header? Really not sure